+++ zerver/tests/test_events.py
@@ -1899,6 +1899,7 @@ class RealmPropertyActionTest(BaseAction):
             default_code_block_language=["python", "javascript"],
             message_content_delete_limit_seconds=[1000, 1100, 1200],
             invite_to_realm_policy=[4, 3, 2, 1],
+            move_messages_between_streams_policy=[4, 3, 2, 1],
         )
 
         vals = test_values.get(name)