+++ zerver/tests/test_home.py
@@ -177,6 +177,7 @@ class HomeTest(ZulipTestCase):
         "realm_message_content_delete_limit_seconds",
         "realm_message_content_edit_limit_seconds",
         "realm_message_retention_days",
+        "realm_move_messages_between_streams_policy",
         "realm_name",
         "realm_name_changes_disabled",
         "realm_name_in_notifications",